Adds support for multiple path parameters
This adds the ability to have multiple path parameters to Beeline router
definitions.

There are several available approaches for doing this. I opted to take
an approach that does not require an additional type parameter to be
specified at the router level. I.E., we can still write

```
appRouter :: Router r => r AppRoute
```

rather than needing to write

```
appRouter :: Router => r AppRoute AppRoute
```

Protecting this top-level context from the extra type parameter
unfortunately required adding a new inner context level where the fully
constructor route type to full values from during generate can vary from
the type's constructor function as the parameters are being specified.

The `end` and `subroute` functions begin construction of a route in this
next context and `route` finishes it when all the parameters of the
constructor function have been filled.

This ends up separating the HTTP method from the route data constructor,
but I think this is probably the most accurate representation of the
router as a tree, where the HTTP methods are the the leaves. The data
constructors for the route values are the trunks the of the branches,
and the route list structure is the joints between the branches.