Fixes sending multiple Accept headers on HTTP requests
Previously we were sending an accept header for each successful response
code. Instead we now ask that the request and response body schemas
define their content-type directly rather than an arbitrary list of
headers so that we manage the `Content-Type` and `Accept` headers more
directly.

We also now include the content types from _all_ the response schemas
rather than just the successful ones. Excluding the non-successful
response codes could cause a server that follows the HTTP Spec to
returna 406 not acceptable in cases where it could not produce the same
content type as was specified for a successful response.